feat : [+] map_renderer acquired engine's renderer [+] fixed bsp traversing nodes functions [+] draws subsectors
[+] basic input control

+++ include/bsp.h
@@ -1,13 +1,14 @@
 #ifndef BSP_H
 #define BSP_H
 
+#include "map_renderer.h"
 #include "structs.h"
 
-#define SUBSECTOR_IDENTIFIER 0x8000
+#define SUBSECTOR_IDENTIFIER (size_t)0x8000
 
 bsp *bsp_init(engine *e, player *p);
 
-void render_bsp_node(bsp *b, int node_id);
+void render_bsp_node(bsp *b, size_t node_id);
 
 void update_bsp(bsp *b);
